% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIZ_alter.cols.R
\name{alter.cols}
\alias{alter.cols}
\title{Modify vectors of colors}
\usage{
alter.cols(x, alpha = NA, mix = NA, wgt = 0.5)
}
\arguments{
\item{x}{A vector of colors specified either by name, hexadecimal code, or numbers (corresponding to colors defined in
\code{palette()}). RGB values can be coerced to hexadecimal codes via the \code{rgb()}.}

\item{alpha}{A numeric vector of alpha (transparency) values, with 0 corresponding to complete transparency and 1 to complete
opaqueness. Use NA to not modify transparency of colors in \code{x}.}

\item{mix}{Another vector of colors to mix with the colors in \code{x}. Use NA to prevent any mixing.}

\item{wgt}{A numeric vector of weights to control how colors in \code{mix} are combined with colors in \code{x}, with 0 
corresponding to no mixing and 1 to complete replacement.}
}
\value{
a character vector of color values in hexadecimal code.
}
\description{
This function alters a vector of colors by modifying their transparency and/or mixing them with other colors.
}
\examples{
colors<-c('red','green','blue')
alter.cols(colors,alpha=c(0,1,0.5),mix=c('white','black','gray'),wgt=c(0,0.5,1))
#recycling behavior--recycled to maximum possible length
alter.cols(colors,alpha=0.2,mod.val=c('black','white'),wgt=c(0.1,0.2,0.8,0.9))

}
