% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mis_covenient.methods.R
\name{anc.edges}
\alias{anc.edges}
\title{Extract indices for edge ancestral to each edge in a phylogeny}
\usage{
anc.edges(phy)
}
\description{
Extract indices for edge ancestral to each edge in a phylogeny
}
