% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contSimmap.R
\docType{package}
\name{contSimmap}
\alias{contSimmap}
\title{contSimmap: A package for relaxed rate Brownian Motion models and continuous stochastic character maps on phylogenies}
\description{
contSimmap more detailed description
}
\section{function catergory 1}{

functions...
}

