% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mis_covenient.methods.R
\name{des.edges}
\alias{des.edges}
\title{Extract indices for edges descending from each edge in a phylogeny}
\usage{
des.edges(phy)
}
\description{
Extract indices for edges descending from each edge in a phylogeny
}
