% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mis_covenient.methods.R
\name{drop.tips}
\alias{drop.tips}
\title{Drop tips from objects with phylogenetic trees and edgewise/nodewise data}
\usage{
drop.tips(phy, tip, partial.match, entire.clade, invert, ...)
}
\arguments{
\item{phy}{An object of class \code{evorates}, \code{evorates_fit}, or \code{phylo}.}

\item{tip}{A numeric or character vector specifying which tips to drop from \code{phy}.}

\item{partial.match}{\code{TRUE} or \code{FALSE}: if \code{tip} is a character vector, should it be
partially matched to tip labels? For example, one could write a genus name to select all
tips belonging to that genus. Defaults to \code{FALSE} to align with normal \link[ape]{drop.tip}
behavior by default.}

\item{entire.clade}{\code{TRUE} or \code{FALSE}: should \emph{all} descendants of the
most recent common ancestor of tips in \code{tip} be removed? Allows for easy exclusion of
entire monophyletic clades. Defaults to \code{FALSE} to align with normal \link[ape]{drop.tip}
behavior by default.}

\item{invert}{\code{TRUE} or \code{FALSE}: should \code{tip} specify tips to keep, rather than
to remove, as in \link[ape]{keep.tip}? Defaults to \code{FALSE} to align with normal
\link[ape]{drop.tip} behavior by default.}

\item{...}{Optional arguments. Notably, arguments from \link[ape]{drop.tip} can only be used
with \code{phylo} objects. In the cases of \code{evorates} and \code{evorates_fit}, arguments
from \link[ape]{drop.tip} are set to their defaults with the exception of \code{collapse.singles},
which is set to \code{FALSE} (though I hope to allow this option to be \code{TRUE} in the future).
Other optional arguments in the case of \code{evorates} and
\code{evorates_fit} objects include:
\itemize{
\item{\code{try.all.elements} (for \code{evorates} objects only), which, if \code{TRUE} (the default),
causes the function to assume any vector with the same number of elements as there are edges in the
phylogeny corresponds to edgewise data and reshuffles indices accordingly. This will additionally sort the
\code{post.probs} element present in most \code{evorates} objects coerced from \code{evorates_fit} objects,
but may cause unpredictable results in cases when users manually customize an \code{evorates} objects.}
\item{\code{store.traits}, which, if \code{TRUE} (the default), causes the function to infer/store trait
values for all nodes in \code{phy}. This is helpful because dropping tips will inevitably alter the
inference of trait values--particularly ancestral state reconstructions--and this is probably undesirable
in most cases.}
\item{\code{recalc.bg} (for \code{evorates_fit} objects only), which, if \code{TRUE} (defaults to \code{FALSE}),
causes the function to recalculate the background rate ("\code{bg_rate}") parameter based on the new
phylogenetic tree, which may be desirable in some cases.}
\item{\code{recalc.devs} (for \code{evorates_fit} objects only), which, if \code{TRUE} (defaults to \code{FALSE}),
causes the function to recalculate rate deviations/posterior probabilities to determine which branchwise
rates are "anomalously" fast or slow based on the new phylogenetic tree, which may be desirable in some cases.
The arguments \code{remove.trend} and \code{geometric} may also be specified to further control of the (re)calculation
of rate deviations (see \link{get.bg.rate} for details).}
}}
}
\value{
An object of the same class as \code{phy}, with reordered phylogeny and
edgewise/nodewise data.
}
\description{
This is a generic that extends the \code{drop.tip()} function from \code{ape} to objects of class
\code{evorates} and \code{evorates_fit}. Mostly intended for dropping "pseudo-tips" used for
fossil calibrations, but could be helpful in a number of scenarios.
}
\seealso{
\link[ape]{drop.tip}, \link{get.bg.rate}
}
