% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mis_covenient.methods.R
\name{edge.ranges}
\alias{edge.ranges}
\title{Extract start and end times for each edge in a phylogeny}
\usage{
edge.ranges(phy)
}
\description{
Extract start and end times for each edge in a phylogeny
}
