% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mis_clade.edges.R
\name{exclude.clade}
\alias{exclude.clade}
\title{Exclude edge indices for a nested subclade out of a larger clade}
\usage{
exclude.clade(
  tree,
  node1 = NULL,
  node2 = NULL,
  return.both = FALSE,
  edge.group1 = NULL,
  edge.group2 = NULL
)
}
\description{
Exclude edge indices for a nested subclade out of a larger clade
}
