% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fastpgram}
\alias{fastpgram}
\title{SimBM' function so I can easily switch between random sampling and this type of sampling when simulating BM bridges}
\usage{
fastpgram(x, tree, add = F, node.col = "black", edge.col = "black",
  ...)
}
\description{
SimBM' function so I can easily switch between random sampling and this type of sampling when simulating BM bridges
}
