% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_operators.R
\name{\%MAPs\%}
\alias{\%MAPs\%}
\title{Extract max a posteriori parameter estimates from a fitted correlated rates Brownian Motion model}
\usage{
fit \%MAPs\% select
}
\arguments{
\item{fit}{An object of class "\code{corateBM_fit}", the output of a call to \code{fit.corateBM}.}

\item{select}{A character or numeric vector. If of class character, the given text is matched to parameter
names using regular expressions. If of class numeric, the numbers are matched to indices of rate parameters
("\code{R_i}"), including the root ("\code{R_0}"). Exhibits special behavior with regards to rate deviation
parameters ("\code{R_i_dev}") and commas ("\code{,}"); see details below.}
}
\value{
A numeric vector or matrix. The dimensions will always go in the order of parameters then chains,
collapsing any dimensions of length 1 and storing associated name information in attributes.
}
\description{
This is an operator for efficiently extracting sampled parameter values from iterations that exhibited the
highest posterior probability for each chain in the output of the function \code{fit.corateBM}. Doesn't
require a pre-stored \code{MAPs} element.
}
\details{
Note that finding MAP estimates is not the goal of Bayesian inference per se. This is especially
true of evolving rates models, in which the posterior probability may be inflated towards infinity by
sampling edgewise rates from multivariate normal distributions with small covariance matrices (to see why,
think about what the probability density of 0 is for a normal distribution with mean and standard deviation
0). The reason this region of parameter space doesn't dominate the posterior is that it's incredibly narrow,
and therefore doesn't meaningfully contribute to the overall posterior probability \emph{integrating over
parameter space}. However, because of this, MAPs will always be associated with tiny \code{R_sig2} values.
This is a feature of just about any hierarchical Bayesian model. Nonetheless, this function is provided for
convenience--it may be useful in situation where \code{R_sig2} is constrained to be 0, as in Brownian Motion
and Early Burst models.


These operators become really useful with some knowledge on how to use regular expressions.
Please consult Rstudio's
\href{https://rstudio.com/wp-content/uploads/2016/09/RegExCheatsheet.pdf}{Regular Expressions Cheat Sheet}
to learn more about regular expressions.

For convenience, when faced with text including a single comma ("\code{,}"), this operator will attempt to
switch the text preceding the comma with any text comeing after the comma and before an underscore 
("\code{_}"). In practice, this helps extract parameters related to covariance matrices, as
"\code{trait1,trait2_evocov}" is automatically converted to "\code{trait2,trait1_evocov}". This is helpful
because corateBM model outputs only store values for the lower triangle of the covariance matrix, since
they are symmetric. Can return unpredictable results in cases where "\code{,}" or "\code{_}" are part of
trait names or tip labels. Use double-backslashes ("\code{\\\\,}" or "\code{\\\\_}") to indicate "escaped"
commas and underscores that the function should ignore during this swapping behavior. Note that you can
also use "\code{\\\\,}" to do encode regular expression searches of the form "\code{{m,n}}".

Since rate deviation parameters are oftentimes redundant with rate parameters, parameter names ending in
"\code{_dev}" are by default excluded unless text with "\code{dev}" is explicitly included at the end of a
text element the select argument.
}
\examples{
#requires example fitted model object
#get rate posterior distributions
example.fit\%MAPs\%1:nrow(example.fit$call$tree$edge)
#',' behavior
example.fit\%MAPs\%'X2,X1_evocov'
#'dev' behavior
example.fit\%MAPs\%'.' #all parameters EXCEPT for rate deviation parameters
example.fit\%MAPs\%c('.','dev') #all parameters


}
\seealso{
Other corateBM_fit operators: 
\code{\link{\%chains\%}()},
\code{\link{\%diagnostics\%}()},
\code{\link{\%means\%}()},
\code{\link{\%quantiles\%}()},
\code{\link{\%sampler\%}()}
}
\concept{corateBM_fit operators}
