% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUT_OPS_main.R
\name{\%means\%}
\alias{\%means\%}
\title{Extract posterior distribution means from a fitted evorates model}
\usage{
fit \%means\% select
}
\description{
Extract posterior distribution means from a fitted evorates model
}
