% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUT_OPS_main.R
\name{\%quantiles\%}
\alias{\%quantiles\%}
\title{Extract posterior distribution quantiles from a fitted evorates model}
\usage{
fit \%quantiles\% select
}
\description{
Extract posterior distribution quantiles from a fitted evorates model
}
