% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT_main.R
\name{input.evorates}
\alias{input.evorates}
\title{Prepare input data for an Evolving Rates model}
\usage{
input.evorates(
  tree,
  trait.data,
  trait.se = NULL,
  constrain.Rsig2 = FALSE,
  trend = FALSE,
  lik.power = 1,
  sampling.scale = FALSE,
  ...
)
}
\arguments{
\item{tree}{An object of class "\code{phylo}"}

\item{trait.data}{Three options:
\itemize{
\item{A named vector of trait values.}
\item{A rownamed matrix of trait values. For the moment, must be 1 column since multivariate models are not
      yet supported.}
\item{A data.frame with 2 columns: 1 with numeric data which is interpreted as trait values and 1 with
      string/factor data interpreted as names. If more than 1 of either kind of column is found, returns error.
      Will also use rownames if no string/factor column is found, but this limits data to consist of only 0-1
      observations per tip.}
}
Both multiple observations for a single tip and missing observations are allowed.
In all cases, the associated names must generally match the tip labels found in \code{tree} (\code{tree$tip.label})
exactly. The only exception are numbers corresponding to internal node indices (these can be viewed by running:
\code{plot(tree); nodelabels()}), which are used to assign trait values to internal nodes. As of now, the function
actually completely ignores \code{tree$node.label}--I should probably fix that and plan to in the future. Be aware
that this feature slightly alters \code{tree} because it works by grafting extra tips of length 0 to the appropriate
nodes.}

\item{trait.se}{A vector, matrix, or data.frame of trait value standard errors which must be unambiguously 
labeled (see \code{trait.data}). Alternatively, a single, unlabeled number that will be applied to all tips in
\code{tree} (e.g., you could set it to 0 to specify all trait values are known without error). If \code{NULL}
(the default), standard error is estimated for all tips in the tree while fitting the model. There are several
things to note here: 
\itemize{
\item{Unlike \code{trait.data}, there can only be 1 trait standard error per tip.}
\item{You can use \code{NA} to specify tips that you want to estimate standard error for and \code{Inf} to
      specify tips that have missing trait values.}
\item{Generally, tips with multiple observations and no observations are automatically assigned standard errors
      of \code{NA} and \code{Inf}, respectively.}
\item{Any tips with unspecified standard error default to \code{NA}.}
}
Any conflicting/impossible standard error specifications are corrected and return warnings.}

\item{constrain.Rsig2}{\code{TRUE} or \code{FALSE}: should the rate varinace (\code{R_sig2}) parameter be constrained to 0,
resulting in a simple Brownian Motion or early/late burst model? Defaults to \code{FALSE}. See Details for a
definition of model parameters.}

\item{trend}{\code{TRUE} or \code{FALSE}: should a trend in rates over time (\code{R_mu}) be estimated,
resulting in an early/late burst or trended evorates model? Defaults to \code{FALSE}.}

\item{lik.power}{A single number between 0 and 1 specifying what power to raise the likelihood function to.
This is useful for sampling from "power posteriors" that shift the posterior to look more like the prior (indeed,
you can set this to 0 to sample from the prior itself). Useful for model diagnostics and calculating things
like Bayes Factors. Technically, you can set \code{lik.power} above 1 (a technique called "data cloning") but
this is not beneficial for evorates models and we don't recommend it.}

\item{sampling.scale}{\code{TRUE} or \code{FALSE}: should provided prior parameters (see \code{...}) be
interpreted on the raw scale of the data or on the transformed scale? All data passed to the Stan-based HMC
sampler are transformed such that \code{tree}'s total height is 1 and the variance of the trait data
is 1. Defaults to \code{FALSE}, such that prior parameters are interpreted on the untransformed scale.}

\item{...}{Prior arguments (see details for further information on what parameters mean):
\itemize{
\item{In general, prior standard deviations and, in some cases, means, can be tweaked by passing arguments
      named "<parameter_name>_sd" and "<parameter_name>_mean", respectively. For example:
      \code{R_mu_mean = 1} or \code{R_sig2_sd = 1}.}
\item{Prior on rate variance (\code{R_sig2}): Follows a half-Cauchy distribution (basically a half-normal distribution 
      with extremely fat tails) with adjustable standard deviation. By default, standard deviation is set to 5 divided
      by the maximum height of \code{tree} (i.e., 5 on the sampling scale).}
\item{Prior on trend (\code{R_mu}): Follows a normal distribution with adjustable mean and standard deviation. By
      default, mean is set to 0 (also 0 on sampling scale), and standard deviation is set to 10 divided by the
      maximum height of \code{tree} (i.e., 10 on the sampling scale).}
\item{Prior on rate at the root (\code{R_0}): Follows a normal distribution with adjustable mean and standard deviation.
      By default, mean is set to variance of \code{trait.data} divided by the maximum height of \code{tree} on the
      natural log scale (i.e., 0 on sampling scale), and standard deviation is set to 10 (also 10 on sampling scale).}
\item{Prior on tip error variance (\code{Y_sig2}): Follows a truncated t distribution (basically a normal
      distribution with potentially fatter tails) with adjustable mean, standard deviation, and degrees of freedom.
      The distribution is truncated at 0 to the left such that Y_sig2 estimates are always positive (note that setting
      the mean to 0 results in a half-t distribution). By default, mean is set to 0, standard deviation is set to 2
      times the variance of \code{trait.data} (i.e., 2 on the sampling scale), and degrees of freedom is set to 1. The
      degrees of freedom can be tweaked with the argument "Ysig2_df" and may be any positive number, including \code{Inf},
      with lower numbers leading to a less informative priors with fatter tails. Notably, setting "Ysig2_df" to \code{Inf}
      makes this prior a truncated normal and much more informative!}
}}
}
\value{
A list of \code{call}, mainly containing information on the inputted tree and trait data, \code{trans.const},
containing transformation constants used to scale data (see \code{sampling.scale}), and \code{dat}, containing the
processed data that will be passed to directly to the Stan-based Hamiltonian Monte Carlo sampler.


#' @details Parameter definitions:
\itemize{
\item{Rate variance (\code{R_sig2}, also denoted \eqn{\sigma^2_{\sigma^2}}{\sigma^2[\sigma^2]}): Determines how much random variation
      accumulates in rates over time. Specifically, independent lineages evolving for a length of time \code{t} would exhibit log-normally
      distributed rates with standard deviation \code{sqrt(t * R_sig2)}. Another way to think about this is that the 95\% credible interval 
      for fold-changes in rates over 1 unit of time is given by \code{exp(c(-1,1) * 1.96 * sqrt(R_sig2))}, assuming \code{R_mu = 0.}}
\item{Trend (\code{R_mu}, also denoted \eqn{\mu_{\sigma^2}}{\mu[\sigma^2]}): Determines whether median rates tend to decrease (if
      negative) or increase (if positive) over time. Specifically, the median fold-change in rate for a given lineage is 
      \code{exp(t * R_mu)} over a length of time \code{t}. This is distinct from changes in average rates, which depends on both
      \code{R_mu} and \code{R_sig2}.}
\item{Note on combining rate variance and trend parameters: The 95\% credible interval 
      for fold-changes in rates over 1 unit of time, given a trend, is simply
      \code{exp(R_mu) * exp(c(-1,1) * 1.96 * sqrt(R_sig2))} or equivalently \code{exp(R_mu + c(-1,1) * 1.96 * sqrt(R_sig2))}.
      This distribution of rate change is right-skewed due to the \code{exp()} function. Because of this, median changes in rates
      over time will always be lower than average changes. The "average change" parameter, denoted \code{R_del} or
      \eqn{\delta_{\sigma^2}}{\delta[\sigma^2]}, is given by \code{R_mu + R_sig2 / 2}. Accordingly, the average fold-change in rate
      for a given lineage is \code{exp(t * R_del)} over a length of time \code{t}.}
\item{Tip error variance (\code{Y_sig2}, also denoted \eqn{\sigma^2_y}{\sigma^2[y]}): Determines the among of "error" around observed
      trait values for tips without fixed standard errors. Specifically, raw observations for a given tip are sampled from a normal
      distribution centered at that tip's "true trait value" with standard deviation \code{sqrt(Y_sig2)}.}
\item{Rate at the root (\code{R_0}, also denoted \eqn{\sigma^2_0}{\sigma^2[0]}): The natural log of the rate at the root of the entire phylogeny.}
\item{Branchwise rates (\code{R_i}, also denoted \eqn{ln \bar{\sigma^2_i}}{ln \sigma^2[i]}, where i is the index of an edge in \code{tree}): 
      The natural log of the average rate along branches of the phylogeny. Note that rates are always shifting over time under this model,
      and these quantities are thus averages. The true rate value at any particular time point along a branch is a related, but separate,
      quantity. These will be NA for branches of length 0.}
\item{Background rate (\code{bg_rate}): The natural log of the average trait evolution rate for the entire phylogeny, given by the average of
      \code{exp(R_i)}, with each entry weighted by its respective branch length. NAs are ignored in this calculation since they correspond
      to branches of length 0.}
\item{Branchwise rate deviations (\code{Rdev_i}, also denoted ln \eqn{\bar{\sigma^2_{dev,i}}}{ln \sigma^2[dev,i]}): Determines whether a
      branches exhibit relatively "fast" (if positive) or "slow" rates (if negative). Generally, these are the differences
      between the branchwise rates and \emph{geometric} background rate on the natural log scale,
      though this depends on \code{remove.trend)}. Here, the geometric background rate is defined as the weighted average of \code{R_i},
      with weights corresponding to branch lengths. This helps prevent some issues with comparing highly right-skewed distributions.
      If \code{remove.trend = TRUE}, then branchwise rates are first "detrended" prior to calculating background rates and deviations
      (\code{R_i - (-log(abs(R_mu)) - log(l_i) + log(abs(exp(R_mu * t1_i) - exp(R_mu * t0_i))))}, where \code{l} is a vector of
      branch lengths and \code{t0}/\code{t1} are vectors of start and end times of each branch). This basically make branchwise rate deviations
      determine whether branches exhibit slow/fast rates \emph{given} the overall trend in rates through time. Otherwise, branchwise rate
      deviations will simmply indicate slow/fast branches occur at the root/tips of a tree in the presence of a strong trend.}
}
}
\description{
This function processes tree and trait data in preparation for fitting an evorates model. Also takes additional model inputs
(e.g., priors), though these may be overwritten later by \code{run.evorates()}.
}
\examples{
#get whale/dolphin tree/trait data
data("cet_fit")
tree <- cet_fit$call$tree
X <- cet_fit$call$trait.data

#prepare data for evorates model
input <- input.evorates(tree = tree, trait.data = trait.data)
#run the evorates model sampler (takes a couple minutes)
run <- run.evorates(input.evorates.obj = input, out.file = "test_fit", chains = 1)
#process the output from the object
fit <- output.evorates(run.evorates.obj = run)
#process the output from the files
fit <- output.evorates(run.evorates.obj = "test_fit")

#see fit.evorates() documentation for further examples


}
\seealso{
\link{fit.evorates} is a convenient wrapper for \link{input.evorates}, \link{run.evorates}, and \link{output.evorates}
}
