% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mis_covenient.methods.R
\name{ladder}
\alias{ladder}
\title{Ladderize objects with phylogenetic trees and edgewise data}
\usage{
ladder(phy, ...)
}
\arguments{
\item{phy}{An object of class \code{evorates}, \code{evorates_fit}, or \code{phylo}.}

\item{...}{Optional arguments including \code{right} and \code{try.all.elements}. When \code{right} is \code{TRUE},
smaller clades are generally sorted to the right/bottom of the phylogeny, and vice versa
when \code{right} is \code{FALSE} (\code{TRUE} by default). \code{try.all.elements} only works with \code{evorates}
objects, and works assuming any vector in the object with the same number of elements as there are edges in the
phylogeny corresponds to edgewise data, thereby sorting it (\code{TRUE} by default). This will additionally sort the
\code{post.probs} element present in most \code{evorates} objects coerced from \code{evorates_fit} objects, but may cause
unpredictable results in cases when users manually customize an \code{evorates} objects.}
}
\value{
An object of the same class as \code{phy}, with reordered phylogeny and edgewise data.
}
\description{
This is a generic that extends the \code{ladderize()} function from \code{ape} to objects of class
\code{evorates} and \code{evorates_fit}. Use it to make tree plots prettier.
}
\examples{
#requires example simulation object
ladder(example.sim)
ladder(example.sim$tree)
#requires example fitted model object
ladder(example.fit)
ladder(example.fit$tree)


}
\seealso{
\link[ape]{ladderize}
}
