% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT_main.R
\name{output.evorates}
\alias{output.evorates}
\title{Process the output of an Evolving Rates model}
\usage{
output.evorates(
  run.evorates.obj,
  stanfit = NULL,
  call = NULL,
  trans.const = NULL,
  dat = NULL,
  include.warmup = FALSE,
  report.quantiles = c(0.025, 0.5, 0.975),
  report.means = TRUE,
  report.devs = TRUE,
  remove.trend = TRUE
)
}
\arguments{
\item{run.evorates.obj}{Either the output of \code{run.evorates()} or a string specifying files to load. These
files always consist of some base name followed by a variable suffix: either "_<X>.csv", where <X> is a chain
number, or "_info". These files contain Stan-based HMC samples/information and auxiliary information related to
data input/transformation constants, respectively. In general, the string should consist of \emph{just} the
base name, not the variable suffix, though the function does attempt to remove suffixes if detected.}

\item{stanfit, call, trans.const, dat}{Generally, you should \emph{not mess with these} unless you know what you're doing!
These are the basic components the function uses to form its output. \code{stanfit} is the Stan-based HMC
samples/information, while the rest make up the auxiliary information regarding input and transformation constants.
You can use these arguments to overwrite certain parts of \code{run.evorates.obj} for debugging purposes.}

\item{include.warmup}{\code{TRUE} or \code{FALSE}: should warmup be included in posterior samples? Warmup is
always included for parameters used to tune the HMC chain, but warmup may be included or excluded for actual
estimated parameters. Defaults to \code{FALSE}.}

\item{report.quantiles}{A vector of posterior distribution quantiles to return (should be between 0 and 1). Set to
\code{NULL} to not return any quantiles. Defaults to 2.5\%, 50\%, and 97.5\% quantiles.}

\item{report.means}{\code{TRUE} or \code{FALSE}: should posterior distribution means be returned? Defaults to
\code{TRUE}.}

\item{report.devs}{\code{TRUE} or \code{FALSE}: should the difference between branchwise rates 
and the overall "background rate" on the natural log scale ("rate deviations") be returned?
The background rate is simply the mean branchwise rate, weighted by each branch's respective length.
If \code{TRUE}, also calculates the posterior probability rate deviations are greater than 0. These
additional parameters help give a sense of which branches in \code{tree} exhibit anomalous trait evolution rates.
Defaults to \code{TRUE}, but is automatically switched to \code{FALSE} when fitting a Brownian
Motion model (\code{constrain.Rsig2 = FALSE & trend = FALSE}).}

\item{remove.trend}{\code{TRUE} or \code{FALSE}: should the rate deviation calculations remove (i.e., "account for") the
effect of trends in rates over time? This may be helpful since strong trends in rates can mask otherwise anomalously
high or low trait evolution rates in certain parts of the tree. Defaults to \code{TRUE}, but has no effect if no trend was
fitted or if \code{report.devs} is \code{FALSE}. If \code{TRUE}, note that \code{report.devs} will be switched to \code{FALSE}
if fitting early/late burst model (\code{constrain.Rsig2 = FALSE & trend = TRUE}).}
}
\value{
An object of class "\code{evorates_fit}" if \code{return.as.obj = TRUE}. Otherwise, the directory
results were saved to (see \code{out.file} for details). An \code{evorates_fit} object is a list of at least
5 components:
\itemize{
\item{\code{call}, which contains information on the final tree, trait values, trait standard errors, and prior
parameters passed to Stan's HMC sampling algorithm (on untransformed scale for better interpretability,
see \code{sampling.scale}).}
\item{\code{sampler.control}, which contains various information on the HMC run, including the number of chains,
iterations, warmup, thinning rate, etc.}
\item{\code{sampler.params}, an array of class "\code{param_block}", containing parameters/diagnostics that were used to tune
the behavior of the HMC while Stan ran it, as well as the (log) prior (\code{prior__}), likelihood (\code{lik__}),
and posterior probability (\code{post__}) of each iteration in the HMC. See Stan manual for more information on
what the parameters mean. The likelihood is not raised to \code{lik.power} here, but the log posterior probability
is calculated while accounting for \code{lik.power}. This always includes warmup iterations, though these can be
discarded using \code{exclude.warmup()} or \code{combine.chains()}.}
\item{\code{param.diags}, a \code{param_block} array, containing diagnostics for each parameter
estimated during the fit, including the
initial value of the HMC chain (\code{init}), the bulk effective sample size (\code{bulk_ess}), the tail
effective sample size (\code{tail_ess}), and the Rhat (\code{Rhat}). See \code{?rstan::Rhat} for more details
on what these diagnostics mean. Generally, you want effective sample sizes to be greater than 400ish and Rhat to be
less than 1.01ish. The functions \code{check.ess()} and \code{check.mix()} will check these thresholds for you
automatically.}
\item{\code{chains}, another \code{param_block} array containing sampled parameter values for each parameter estimated during the fit. See
details for further information on what each parameter means.}
\item{The object optionally contains more \code{param_block} arrays of posterior distribution quantiles (\code{quantiles})
and means (\code{means}), and posterior probabilities (\code{post.prob}, see \code{report.devs} and \code{remove.trend}).}
}
All \code{param_block} arrays' dimensions go in the order of iterations/diagnostics/quantiles, then parameters, then chains.
}
\description{
This function processes raw Stan-based Hamiltonian Monte Carlo (HMC) samples and associated information and returns
all this information in a (relatively) user-friendly format.
}
\details{
Parameter definitions:
\itemize{
\item{Rate variance (\code{R_sig2}, also denoted \eqn{\sigma^2_{\sigma^2}}{\sigma^2[\sigma^2]}): Determines how much random variation
      accumulates in rates over time. Specifically, independent lineages evolving for a length of time \code{t} would exhibit log-normally
      distributed rates with standard deviation \code{sqrt(t * R_sig2)}. Another way to think about this is that the 95\% credible interval 
      for fold-changes in rates over 1 unit of time is given by \code{exp(c(-1,1) * 1.96 * sqrt(R_sig2))}, assuming \code{R_mu = 0.}}
\item{Trend (\code{R_mu}, also denoted \eqn{\mu_{\sigma^2}}{\mu[\sigma^2]}): Determines whether median rates tend to decrease (if
      negative) or increase (if positive) over time. Specifically, the median fold-change in rate for a given lineage is 
      \code{exp(t * R_mu)} over a length of time \code{t}. This is distinct from changes in average rates, which depends on both
      \code{R_mu} and \code{R_sig2}.}
\item{Note on combining rate variance and trend parameters: The 95\% credible interval 
      for fold-changes in rates over 1 unit of time, given a trend, is simply
      \code{exp(R_mu) * exp(c(-1,1) * 1.96 * sqrt(R_sig2))} or equivalently \code{exp(R_mu + c(-1,1) * 1.96 * sqrt(R_sig2))}.
      This distribution of rate change is right-skewed due to the \code{exp()} function. Because of this, median changes in rates
      over time will always be lower than average changes. The "average change" parameter, denoted \code{R_del} or
      \eqn{\delta_{\sigma^2}}{\delta[\sigma^2]}, is given by \code{R_mu + R_sig2 / 2}. Accordingly, the average fold-change in rate
      for a given lineage is \code{exp(t * R_del)} over a length of time \code{t}.}
\item{Tip error variance (\code{Y_sig2}, also denoted \eqn{\sigma^2_y}{\sigma^2[y]}): Determines the among of "error" around observed
      trait values for tips without fixed standard errors. Specifically, raw observations for a given tip are sampled from a normal
      distribution centered at that tip's "true trait value" with standard deviation \code{sqrt(Y_sig2)}.}
\item{Rate at the root (\code{R_0}, also denoted \eqn{\sigma^2_0}{\sigma^2[0]}): The natural log of the rate at the root of the entire phylogeny.}
\item{Branchwise rates (\code{R_i}, also denoted \eqn{ln \bar{\sigma^2_i}}{ln \sigma^2[i]}, where i is the index of an edge in \code{tree}): 
      The natural log of the average rate along branches of the phylogeny. Note that rates are always shifting over time under this model,
      and these quantities are thus averages. The true rate value at any particular time point along a branch is a related, but separate,
      quantity. These will be NA for branches of length 0.}
\item{Background rate (\code{bg_rate}): The natural log of the average trait evolution rate for the entire phylogeny, given by the average of
      \code{exp(R_i)}, with each entry weighted by its respective branch length. NAs are ignored in this calculation since they correspond
      to branches of length 0.}
\item{Branchwise rate deviations (\code{Rdev_i}, also denoted ln \eqn{\bar{\sigma^2_{dev,i}}}{ln \sigma^2[dev,i]}): Determines whether a
      branches exhibit relatively "fast" (if positive) or "slow" rates (if negative). Generally, these are the differences
      between the branchwise rates and \emph{geometric} background rate on the natural log scale,
      though this depends on \code{remove.trend)}. Here, the geometric background rate is defined as the weighted average of \code{R_i},
      with weights corresponding to branch lengths. This helps prevent some issues with comparing highly right-skewed distributions.
      If \code{remove.trend = TRUE}, then branchwise rates are first "detrended" prior to calculating background rates and deviations
      (\code{R_i - (-log(abs(R_mu)) - log(l_i) + log(abs(exp(R_mu * t1_i) - exp(R_mu * t0_i))))}, where \code{l} is a vector of
      branch lengths and \code{t0}/\code{t1} are vectors of start and end times of each branch). This basically make branchwise rate deviations
      determine whether branches exhibit slow/fast rates \emph{given} the overall trend in rates through time. Otherwise, branchwise rate
      deviations will simmply indicate slow/fast branches occur at the root/tips of a tree in the presence of a strong trend.}
}
}
\examples{
#get whale/dolphin tree/trait data
data("cet_fit")
tree <- cet_fit$call$tree
X <- cet_fit$call$trait.data

#prepare data for evorates model
input <- input.evorates(tree = tree, trait.data = trait.data)
#run the evorates model sampler (takes a couple minutes)
run <- run.evorates(input.evorates.obj = input, out.file = "test_fit", chains = 1)
#process the output from the object
fit <- output.evorates(run.evorates.obj = run)
#process the output from the files
fit <- output.evorates(run.evorates.obj = "test_fit")

#see fit.evorates() documentation for further examples


}
\seealso{
\link{fit.evorates} is a convenient wrapper for \link{input.evorates}, \link{run.evorates}, and \link{output.evorates}
}
