% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIM_plot.R
\name{plot.evorates}
\alias{plot.evorates}
\title{Plot simulated trait and/or rate data from a}
\usage{
\method{plot}{evorates}(
  sim,
  traits = 1:ncol(sim$X),
  style = c("phenogram", "phylogram", "cladogram", "fan", "unrooted", "radial"),
  col = def.color.scheme(),
  na.col = "gray90",
  val.range = NULL,
  res = 100,
  alpha = NA,
  breaks = NULL,
  colvec = NULL,
  lwd = 1,
  lty = 1,
  xlab = NULL,
  ylab = NULL,
  add = F,
  color.element = "R",
  exp = FALSE,
  exp.txt = TRUE,
  ...,
  legend = T,
  legend.args = NULL
)
}
\description{
Plot simulated trait and/or rate data from a
}
