% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT_plot.R
\name{plot.evorates_fit}
\alias{plot.evorates_fit}
\title{Plot a fitted evorates model}
\usage{
\method{plot}{evorates_fit}(
  fit,
  chain = NULL,
  type = c("quantiles", "chains", "means", "diagnostics"),
  extra.select = NULL,
  post.probs = c("stored", "recalculate", "none"),
  devs = c("none", "stored", "recalculate"),
  remove.trend = TRUE,
  geometric = TRUE,
  recon.type = NULL,
  recon.extra.select = NULL,
  ...,
  post.probs.args = NULL,
  sim = NULL
)
}
\description{
This function can plot the results of a fitted evorates model in several different styles
}
