% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT_main.R
\name{run.evorates}
\alias{run.evorates}
\title{Run sampler for an Evolving Rates model}
\usage{
run.evorates(
  input.evorates.obj,
  return.as.obj = TRUE,
  out.file = NULL,
  check.overwrite = TRUE,
  constrain.Rsig2 = FALSE,
  trend = FALSE,
  lik.power = 1,
  ...
)
}
\arguments{
\item{input.evorates.obj}{The output of \code{input.evorates()}}

\item{return.as.obj}{\code{TRUE}or \code{FALSE}: should results be passed back to the R
environment as a \code{evorates_fit} object? Defaults to \code{TRUE}. If \code{FALSE}, results are saved to files
with automatically generated names instead (see below).}

\item{out.file}{A directory to save results to. If unspecified, an automatic directory is
generated based on the current date and time and R's working directory. The function will generate csv files
for each chain of the HMC sampler using Stan's built-in functionality (note that these will thus be on the
transformed scale), as well as a separate RDS file giving additional information about the data. Defaults to
\code{NULL}, which means no results are saved to file, though this will be changed automatically if
\code{return.as.obj = FALSE}.}

\item{check.overwrite}{\code{TRUE} or \code{FALSE}: should files in the directory specified by \code{out.file}
be checked to
prevent accidentally overwriting existing files? This part of the code is not thoroughly tested and might take
a long time for folders with many files, so some users may wish to just switch it off. Defaults to \code{TRUE}.}

\item{constrain.Rsig2}{\code{TRUE} or \code{FALSE}: should the rate varinace (\code{R_sig2}) parameter be constrained to 0,
resulting in a simple Brownian Motion or early/late burst model? Defaults to \code{FALSE}. See Details for a
definition of model parameters.}

\item{trend}{\code{TRUE} or \code{FALSE}: should a trend in rates over time (\code{R_mu}) be estimated,
resulting in an early/late burst or trended evorates model? Defaults to \code{FALSE}.}

\item{lik.power}{A single number between 0 and 1 specifying what power to raise the likelihood function to.
This is useful for sampling from "power posteriors" that shift the posterior to look more like the prior (indeed,
you can set this to 0 to sample from the prior itself). Useful for model diagnostics and calculating things
like Bayes Factors. Technically, you can set \code{lik.power} above 1 (a technique called "data cloning") but
this is not beneficial for evorates models and we don't recommend it.}

\item{...}{Other optional arguments:
\itemize{
\item{Prior arguments (see details for further information on what parameters mean):
\itemize{
\item{In general, prior standard deviations and, in some cases, means, can be tweaked by passing arguments
      named "<parameter_name>_sd" and "<parameter_name>_mean", respectively. For example:
      \code{R_mu_mean = 1} or \code{R_sig2_sd = 1}. Note that all prior arguments are necessarily on the
      sampling scale when passed to \code{run.evorates()}}
\item{Prior on rate variance (\code{R_sig2}): Follows a half-Cauchy distribution (basically a half-normal distribution 
      with extremely fat tails) with adjustable standard deviation. By default, standard deviation is set to 5 divided
      by the maximum height of \code{tree} (i.e., 5 on the sampling scale).}
\item{Prior on trend (\code{R_mu}): Follows a normal distribution with adjustable mean and standard deviation. By
      default, mean is set to 0 (also 0 on sampling scale), and standard deviation is set to 10 divided by the
      maximum height of \code{tree} (i.e., 10 on the sampling scale).}
\item{Prior on rate at the root (\code{R_0}): Follows a normal distribution with adjustable mean and standard deviation.
      By default, mean is set to variance of \code{trait.data} divided by the maximum height of \code{tree} on the
      natural log scale (i.e., 0 on sampling scale), and standard deviation is set to 10 (also 10 on sampling scale).}
\item{Prior on tip error variance (\code{Y_sig2}): Follows a half-t distribution (basically a half-normal
      distribution with potentially fatter tails) with adjustable standard deviation and degrees of freedom. By default,
      standard deviation is set to 2 times the variance \code{trait.data} (i.e., 2 on the sampling scale), and
      degrees of freedom is set to 1. The degrees of freedom can be tweaked with the argument "Ysig2_df" and may be
      any positive number, including \code{Inf}, with lower numbers leading to a less informative priors with fatter
      tails. Notably, setting "Ysig2_df" to \code{Inf} makes this prior half-normal and much more informative!
      More informative priors are useful in cases where a user lacks precise estimates of tip error, but nonetheless
      knows that phenotypic divergence is generally higher across, rather than within, tips.}
}}
\item{Additional arguments to pass to \code{rstan::sampling()}, most commonly:
\itemize{
\item{\code{chains} to specify the number of HMC chains (defaults to 4)}
\item{\code{iter} to  specify the number of iterations in HMC chains (defaults to 2000)}
\item{\code{warmup} to specify the number of warmup iterations (defaults to \code{floor(iter/2)})}
\item{\code{thin} to specify which iterations to keep in results (defaults to 1 or no thinning)}
\item{\code{cores} to specify the number of computer cores to use (defaults to \code{getOption("mc.cores", 1L)})}
\item{\code{refresh} to control when progress is reported (defaults to \code{max(iter/10, 1)}, and can be
suppressed by setting to 0 or less)}
\item{There are other things users might want to mess with, like \code{seed}, \code{init}, and \code{control}.
See \code{?rstan::sampling} and \code{?rstan::stan} for further details.}
}}
}}
}
\value{
A list if \code{return.as.obj = TRUE}, containing the unprocessed model fit in \code{stanfit} format and
updated \code{input.evorates.obj}. Otherwise, the directory results were saved to (see \code{out.file} for details).


#' @details Parameter definitions:
\itemize{
\item{Rate variance (\code{R_sig2}, also denoted \eqn{\sigma^2_{\sigma^2}}{\sigma^2[\sigma^2]}): Determines how much random variation
      accumulates in rates over time. Specifically, independent lineages evolving for a length of time \code{t} would exhibit log-normally
      distributed rates with standard deviation \code{sqrt(t * R_sig2)}. Another way to think about this is that the 95\% credible interval 
      for fold-changes in rates over 1 unit of time is given by \code{exp(c(-1,1) * 1.96 * sqrt(R_sig2))}, assuming \code{R_mu = 0.}}
\item{Trend (\code{R_mu}, also denoted \eqn{\mu_{\sigma^2}}{\mu[\sigma^2]}): Determines whether median rates tend to decrease (if
      negative) or increase (if positive) over time. Specifically, the median fold-change in rate for a given lineage is 
      \code{exp(t * R_mu)} over a length of time \code{t}. This is distinct from changes in average rates, which depends on both
      \code{R_mu} and \code{R_sig2}.}
\item{Note on combining rate variance and trend parameters: The 95\% credible interval 
      for fold-changes in rates over 1 unit of time, given a trend, is simply
      \code{exp(R_mu) * exp(c(-1,1) * 1.96 * sqrt(R_sig2))} or equivalently \code{exp(R_mu + c(-1,1) * 1.96 * sqrt(R_sig2))}.
      This distribution of rate change is right-skewed due to the \code{exp()} function. Because of this, median changes in rates
      over time will always be lower than average changes. The "average change" parameter, denoted \code{R_del} or
      \eqn{\delta_{\sigma^2}}{\delta[\sigma^2]}, is given by \code{R_mu + R_sig2 / 2}. Accordingly, the average fold-change in rate
      for a given lineage is \code{exp(t * R_del)} over a length of time \code{t}.}
\item{Tip error variance (\code{Y_sig2}, also denoted \eqn{\sigma^2_y}{\sigma^2[y]}): Determines the among of "error" around observed
      trait values for tips without fixed standard errors. Specifically, raw observations for a given tip are sampled from a normal
      distribution centered at that tip's "true trait value" with standard deviation \code{sqrt(Y_sig2)}.}
\item{Rate at the root (\code{R_0}, also denoted \eqn{\sigma^2_0}{\sigma^2[0]}): The natural log of the rate at the root of the entire phylogeny.}
\item{Branchwise rates (\code{R_i}, also denoted \eqn{ln \bar{\sigma^2_i}}{ln \sigma^2[i]}, where i is the index of an edge in \code{tree}): 
      The natural log of the average rate along branches of the phylogeny. Note that rates are always shifting over time under this model,
      and these quantities are thus averages. The true rate value at any particular time point along a branch is a related, but separate,
      quantity. These will be NA for branches of length 0.}
\item{Background rate (\code{bg_rate}): The natural log of the average trait evolution rate for the entire phylogeny, given by the average of
      \code{exp(R_i)}, with each entry weighted by its respective branch length. NAs are ignored in this calculation since they correspond
      to branches of length 0.}
\item{Branchwise rate deviations (\code{Rdev_i}, also denoted ln \eqn{\bar{\sigma^2_{dev,i}}}{ln \sigma^2[dev,i]}): Determines whether a
      branches exhibit relatively "fast" (if positive) or "slow" rates (if negative). Generally, these are the differences
      between the branchwise rates and \emph{geometric} background rate on the natural log scale,
      though this depends on \code{remove.trend)}. Here, the geometric background rate is defined as the weighted average of \code{R_i},
      with weights corresponding to branch lengths. This helps prevent some issues with comparing highly right-skewed distributions.
      If \code{remove.trend = TRUE}, then branchwise rates are first "detrended" prior to calculating background rates and deviations
      (\code{R_i - (-log(abs(R_mu)) - log(l_i) + log(abs(exp(R_mu * t1_i) - exp(R_mu * t0_i))))}, where \code{l} is a vector of
      branch lengths and \code{t0}/\code{t1} are vectors of start and end times of each branch). This basically make branchwise rate deviations
      determine whether branches exhibit slow/fast rates \emph{given} the overall trend in rates through time. Otherwise, branchwise rate
      deviations will simmply indicate slow/fast branches occur at the root/tips of a tree in the presence of a strong trend.}
}
}
\description{
This function runs a Stan-based Hamiltonian Monte Carlo (HMC) sampler given input data and priors, but only
stores, rather than processes, the output.
}
\examples{
#get whale/dolphin tree/trait data
data("cet_fit")
tree <- cet_fit$call$tree
trait.data <- cet_fit$call$trait.data

#prepare data for evorates model
input <- input.evorates(tree = tree, trait.data = trait.data)
#run the evorates model sampler (takes a couple minutes)
run <- run.evorates(input.evorates.obj = input, out.file = "test_fit", chains = 1)
#process the output from the object
fit <- output.evorates(run.evorates.obj = run)
#process the output from the files
fit <- output.evorates(run.evorates.obj = "test_fit")

#see fit.evorates() documentation for further examples


}
\seealso{
\link{fit.evorates} is a convenient wrapper for \link{input.evorates}, \link{run.evorates}, and \link{output.evorates}
}
