% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIM_generation.R
\name{sim.evorates}
\alias{sim.evorates}
\title{Simulate data under an evorates model}
\usage{
sim.evorates(
  tree,
  R0 = 0,
  Rsig2 = 1,
  X0 = 0,
  Rmu = 0,
  Xsig2 = 1,
  trait.names = NULL,
  n.obs = rep(1, length(tree$tip.label)),
  Ysig2 = 0,
  percent = FALSE,
  anc.states = FALSE,
  slow = FALSE,
  res = 500
)
}
\arguments{
\item{tree}{An object of class "\code{phylo}".}

\item{R0}{The starting rate at the root of the phylogeny (on natural log scale!).}

\item{Rsig2}{The rate variance parameter, which controls the magnitude of random
changes in rates over time.}

\item{Rmu}{The trend parameter, which controls whether rates tend to decrease
or increase over time.}

\item{Xsig2}{You probably shouldn't alter this unless you know what}
}
\description{
This function simulates trait data evolving on a given phylogeny under a model
whereby rates also gradually "evolve" over time.
}
