% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mis_covenient.methods.R
\name{sis.edges}
\alias{sis.edges}
\title{Extract indices for edges sister to each edge in a phylogeny}
\usage{
sis.edges(phy)
}
\description{
Extract indices for edges sister to each edge in a phylogeny
}
