% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_chain.management.R
\name{thin.chains}
\alias{thin.chains}
\title{Thin chains in a fitted correlated rates Brownian Motion model}
\usage{
thin.chains(fit, thin = 2)
}
\arguments{
\item{fit}{An object of class "\code{corateBM_fit}", the output of a call to \code{fit.corateBM}.}

\item{thin}{}
}
\value{
an object of class "\code{corateBM_fit}". All previously-existing elements in \code{fit} will be
included.
}
\description{
This function thins all chains from the output of the function \code{fit.corateBM} by keeping only every
\code{thin}th iteration.
}
\details{
Keep in \code{warmup} counts form the first iteration. For example, if \code{fit} has 2000
iterations with the first 1000 as warmup, you would set \code{warmup} to 1500 to keep the last 500
iterations.
}
\seealso{
Other chain management: 
\code{\link{combine.chains}()},
\code{\link{exclude.warmup}()},
\code{\link{select.chains}()}
}
\concept{chain management}
