% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mis_covenient.methods.R
\name{tip.edges}
\alias{tip.edges}
\title{Extract indices for edges corresponding to tips in a phylogeny}
\usage{
tip.edges(phy, include.names = TRUE)
}
\description{
Extract indices for edges corresponding to tips in a phylogeny
}
