% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mis_alter.cols.R
\name{alter.cols}
\alias{alter.cols}
\title{Modify the transparency and shade of colors}
\usage{
alter.cols(x, alpha = NA, mod.val = 0)
}
\arguments{
\item{x}{Vector of color values, which can either be a character vector of color names and/or hexadecimal codes, or a numeric
vector. In the latter case, the vector will be coerced to integers taken to correspond to colors currently defined in palette().}

\item{alpha}{Vector of alpha (transparency) values (numeric), with 0 corresponding to complete transparency and 1 to complete
opaqueness. Use NA to not modify transparency of original color vector. alph must consist of NA's and/or values between 0 and 1.}

\item{mod.val}{Vector of values (numeric) by which to brighten/darken colors, with positive and negative numbers corresponding to
brightening and darkening, respectively. In this context, rgb values range from 0 to 1, and any modified values exceeding
these boundaries will be rounded up or down, respectively.}
}
\value{
a vector of color values in hexadecimal code (character)
}
\description{
This function takes a vector of colors (x) as input and returns a vector of colors of the same length. The function can be used to
either make the colors a certain transparency (alph) or make the colors lighter or darker (mod.val). alph and mod.val are recycled
or truncated to be the same length as x.
}
\examples{
colors<-c('red','green','blue')
alter.cols(colors,alpha=c(0,1,0.5),mod.val=c(0,-0.5,-0.75))
#recycling behavior 
alter.cols(colors,alpha=0.2,mod.val=-0.5)

}
