% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evorates.R
\docType{package}
\name{evorates}
\alias{evorates}
\title{evorates: Fitting continuously variable rate models to comparative data on continuous traits}
\description{
evorates more detailed description
}
\section{function category 1}{

functions...
}

