% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_chain.management.R
\name{exclude.warmup}
\alias{exclude.warmup}
\title{Exclude warmup in a fitted evolving rates model}
\usage{
exclude.warmup(fit, warmup = fit$sampler.control$warmup, sampler = T)
}
\arguments{
\item{fit}{An object of class "\code{evorates_fit}", the output of a call to \code{fit.evorates}.}

\item{warmup}{the number of iterations to be excluded, counting from the 1st iteration, even if it isn't
stored in \code{fit}.}

\item{sampler}{TRUE or FALSE value: should the iterations be excluded from the \code{sampler.params} element
too?}
}
\value{
an object of class "\code{evorates_fit}". All previously-existing elements in \code{fit} will be
included.
}
\description{
This function gets rid of an arbitrary number of iterations at the beginning of all chains from the output
of the function \code{fit.evorates}.
}
\details{
Keep in \code{warmup} counts form the first iteration. For example, if \code{fit} has 2000
iterations with the first 1000 as warmup, you would set \code{warmup} to 1500 to keep the last 500
iterations.
}
\examples{
#requires example fitted model object
#exclude warmup iterations from sampler
exclude.warmup(example.fit)
#exclude some non-warmup iterations
exclude.warmup(example.fit,1100)
#exclude some non-warmup iterations, but keep in sampler.params
exclude.warmup(example.fit,1100,F)


}
\seealso{
Other chain management: 
\code{\link{combine.chains}()},
\code{\link{select.chains}()},
\code{\link{thin.chains}()}
}
\concept{chain management}
