% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.corateBM.R
\name{fit.corateBM}
\alias{fit.corateBM}
\title{Fit an autocorrelated rate Brownian motion model}
\usage{
fit.corateBM(
  tree,
  X,
  R0.prior = 10,
  Rsig2.prior = 20,
  X0.prior = 100,
  report.quantiles = c(0.025, 0.5, 0.975),
  report.MAPs = T,
  report.devs = T,
  ...
)
}
\arguments{
\item{tree}{A phylogenetic tree of class 'phylo'; function does not yet support multiple phylogenies
but will handle polytomous trees just fine.}

\item{X}{A vector of trait data, with each element labelled according to its corresponding tip label.}

\item{R0.prior}{The standard deviation of the cauchy prior for the \code{R0} parameter}

\item{Rsig2.prior}{The standard deviation of the cauchy prior for the \code{Rsig2} parameter.}

\item{X0.prior}{The standard deviation of the normal prior for the \code{X0} parameter.}

\item{report.quantiles}{A vector of values between 0 and 1, telling the function which quantiles of
of the marginal posterior distributions for each parameter to calculate. Use this to extract
'credible intervals' for parameter values. Set to NULL if no quantiles are desired.}

\item{report.MAPs}{TRUE or FALSE value: should the function calculate mean a posteriori parameter
estimates?}

\item{report.devs}{TRUE or FALSE value: should the function calculate differences between branch-wise
rates and the 'background rate' (see below)?}

\item{...}{Extra arguments to pass to the Stan's mcmc sampler. Use \code{chains} to set the number
of chains to run (4 by default), \code{iter} for the number of iterations (2000 by default),
\code{warmup} for the number of iterations devoted to the warmup period (\code{iter/2} by default),
\code{thin} to set the thinning rate (1 by default), and \code{refresh} to set the frequency of
reporting in the R console (\code{iter/10} by default). See \code{rstan::sampling} for more options.}
}
\value{

}
\description{
This
}
