% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_fun.R
\name{fit.evorates}
\alias{fit.evorates}
\title{Fit an Evolving Rates model}
\usage{
fit.evorates(
  tree,
  trait.data,
  trait.se = NULL,
  constrain.Rsig2 = FALSE,
  trend = FALSE,
  lik.power = 1,
  sampling.scale = FALSE,
  return.as.obj = TRUE,
  out.file = NULL,
  check.overwrite = TRUE,
  include.warmup = FALSE,
  report.quantiles = c(0.025, 0.5, 0.975),
  report.means = TRUE,
  report.devs = TRUE,
  remove.trend = TRUE,
  report.MAPs = FALSE,
  ...
)
}
\arguments{
\item{tree}{An object of class "\code{phylo}"}

\item{trait.data}{Three options:
\itemize{
\item{A named vector of trait values.}
\item{A rownamed matrix of trait values. For the moment, must be 1 column since multivariate models are not
yet supported.}
\item{A data.frame with 2 columns: 1 with numeric data which is interpreted as trait values and 1 with
string/factor data interpreted as names. If more than 1 of either kind of column is found, returns error.
Will also use rownames if no string/factor column is found, but this limits data to consist of only 0-1
observations per tip.}
}
In all cases, the associated names must match the tip labels found in \code{tree} (\code{tree$tip.label})
exactly. Both multiple observations for a single tip and missing observations are allowed.}

\item{trait.se}{A vector, matrix, or data.frame of trait value standard errors which must be unambiguously 
labeled (see \code{trait.data}). Alternatively, a single, unlabeled number that will be applied to all tips in
\code{tree} (e.g., you could set it to 0 to specify all trait values are known without error). If \code{NULL}
(the default), standard error is estimated for all tips in the tree while fitting the model. There are several
things to note here: 
\itemize{
\item{unlike \code{trait.data}, there can only be 1 trait standard error per tip.}
\item{you can use \code{NA} to specify tips that you want to estimate standard error for and \code{Inf} to
specify tips that have missing trait values.}
\item{Generally, tips with multiple observations and no observations are automatically assigned standard errors
of \code{NA} and \code{Inf}, respectively.}
\item{Any tips with unspecified standard error default to \code{NA}.}
}
Any conflicting/impossible standard error specifications are corrected and return warnings.}

\item{constrain.Rsig2}{\code{TRUE} or \code{FALSE}: should the \code{R_sig2} parameter be constrained to 0,
resulting in a simple Brownian Motion or Early/Late Burst model? Defaults to \code{FALSE}. See Details for a
definition of model parameters.}

\item{trend}{\code{TRUE} or \code{FALSE}: should a trend in rates over time (\code{R_mu})  be estimated,
resulting in an Early/Late Burst or trended evorates model? Defaults to \code{FALSE}.}

\item{lik.power}{A single number between 0 and 1 specifying what power to raise the likelihood function to.
This is useful for sampling from "power posteriors" that shift the posterior to look more like the prior (indeed,
you can set this to 0 to sample from the prior itself). Useful for model diagnostics and calculating things
like Bayes Factors. Technically, you can set \code{lik.power} above 1 (a technique called "data cloning") but
this is not beneficial for evorates models and we don't recommend it.}

\item{sampling.scale}{\code{TRUE} or \code{FALSE}: should provided prior parameters (see \code{...}) be
interpreted on the raw scale of the data or on the transformed scale? All data passed to the Stan-based HMC
sampler are transformed such that \code{tree}'s total height is 1 and the standard deviation of the trait data
is 1. Defaults to \code{FALSE}, such prior parameters are interpreted on the untransformed scale.}

\item{return.as.obj}{\code{TRUE}or \code{FALSE}: should results be passed back to the R
environment as a \code{evorates_fit} object? Defaults to \code{TRUE}. If \code{FALSE}, results are saved to files
instead with an automatically generated names (see below).}

\item{out.file}{A directory to save results to. If unspecified, an automatic directory is
generated based on the current date and time and R's working directory. The function will generate csv files
for each chain of the HMC sampler using Stan's built-in functionality (note that these will thus be on the
transformed scale), as well as a separate RDS file giving additional information about the data. Defaults to
\code{NULL}, which means no results are saved to file, though this will be changed automatically if
\code{return.as.obj = FALSE}.}

\item{check.overwrite}{\code{TRUE} or \code{FALSE}: should files in the directory specified by \code{out.file}
be checked to
prevent accidentally overwriting existing files? This part of the code is not thoroughly tested and might take
a long time for folders with many files, so some users may wish to just switch it off. Defaults to \code{TRUE}.}

\item{include.warmup}{\code{TRUE} or \code{FALSE}: should warmup be included in posterior samples? Warmup is
always included for parameters used to tune the HMC chain, but warmup may be included or excluded for actual
estimated parameters. Defaults to \code{FALSE}.}

\item{report.quantiles}{A vecotr posterior distribution quantiles to return (should be between 0 and 1). Set to
\code{NULL} to not return any quantiles. Defaults to 2.5\%, 50\%, and 97.5\% quantiles.}

\item{report.means}{\code{TRUE} or \code{FALSE}: should posterior distribution means be returned? Defaults to
\code{TRUE}.}

\item{report.devs}{\code{TRUE} or \code{FALSE}: should the difference between time-averaged rates along each
branch of \code{tree} and the overall average rate on the natural log scale ("rate deviations") be returned?
If \code{TRUE}, also
calculates the posterior probability a particular time-averaged rate is greater than the overall average. These
additional parameters help give a sense of which branches in \code{tree} exhibit anomalous trait evolution rates
under the model. Defaults to \code{TRUE}, but is automatically switched to \code{FALSE} when fitting a Brownian
Motion model (\code{constrain.Rsig2 = FALSE & trend = FALSE}).}

\item{remove.trend}{\code{TRUE} or \code{FALSE}: should the rate deviation calculations remove (i.e., "account for") the
effect of trends in rates over time? This is sometimes helpful since strong trends in rates can mask otherwise anomalously
high or low trait evolution rates in certain parts of the tree. Defaults to \code{TRUE}, but has no effect if no trend was
fitted or if \code{report.devs} is \code{FALSE}.}

\item{report.MAPs}{\code{TRUE} or \code{FALSE}: should maximum a posteriori parameter estimates be returned?
Defaults to \code{FALSE}.}

\item{...}{Other optional arguments:
\itemize{
\item{Prior arguments: priors on \code{R_0} and \code{R_sig2} follow normal distributions, while priors on \code{R_sig2}
and \code{Y_sig2} follow half-Cauchy distributions, which are basically half-normal distributions with extremely fat tails.
Both the mean and standard deviation of \code{R_0}/\code{R_sig2} priors can be tweaked, while only the standard deviation of
\code{R_sig2}/\code{Y_sig2} can be tweaked. See Details for definitions of what these parameters
mean. To specify a prior mean, pass an argument named "\code{<parameter name>_mean}" (e.g., "\code{R_0_mean}"),
and to specify a prior standard deviation, pass an argument named "\code{<parameter name>_sd}" (e.g.,
"\code{R_mu_sd}").}
\item{Additional arguments to pass to \code{rstan::sampling()}, most commonly:
\itemize{
\item{\code{chains} to specify the number of HMC chains (defaults to 4)}
\item{\code{iter} to  specify the number of iterations in HMC chains (defaults to 2000)}
\item{\code{warmup} to specify the number of warmup iterations (defaults to \code{floor(iter/2)})}
\item{\code{thin} to specify which iterations to keep in results (defaults to 1 or no thinning)}
\item{\code{cores} to specify the number of computer cores to use (defaults to \code{getOption("mc.cores", 1L)})}
\item{\code{refresh} to control when progress is reported (defaults to \code{max(iter/10, 1)}, and can be
suppressed by setting to 0 or less)}
\item{There are other things users might want to mess with, like \code{seed}, \code{init}, and \code{control}.
See \code{?rstan::sampling} and \code{?rstan::stan} for further details.}
}}
}}
}
\value{
An object of class "\code{evorates_fit}" if \code{return.as.obj = TRUE}. Otherwise, nothing, as results
are saved to file instead (see \code{out.file} for details). An \code{evorates_fit} object is a list of at least
5 components:
\itemize{
\item{\code{call}, which contains information on the final tree, trait values, trait standard errors, and prior
parameters passed to Stan's HMC sampling algorithm (on untransformed scale for better interpretability,
see \code{sampling.scale}).}
\item{\code{sampler.control}, which contains various information on the HMC run, including the number of chains,
iterations, warmup, thinning rate, etc.}
\item{\code{sampler.params}, an array of parameters/diagnostics that were used to tune the behavior of the HMC
while Stan ran it, as well as the (log) prior (\code{prior}), likelihood (\code{lik}), and posterior probability
(\code{post}) of each iteration in the HMC. See Stan manual for more information on what the parameters mean.
The likelihood is not raised to \code{lik.power} here, but the log posterior probability is calculated while
accounting for \code{lik.power}. This always includes warmup iterations, though these can be discarded using
\code{exclude.warmup()} or \code{combine.chains()}.}
\item{\code{param.diags}, an array of diagnostics for each parameter estimated during the fit, including the
initial value of the HMC chain (\code{init}), the bulk effective sample size (\code{bulk_ess}), the tail
effective sample size (\code{tail_ess}), and the Rhat (\code{Rhat}). See \code{?rstan::Rhat} for more details
on what these diagnostics mean. Generally, you want effective sample sizes to be greater than 100 and Rhat to be
less than 1.05.}
\item{\code{chains}, an array of sampled parameter values for each parameter estimated during the fit. See
details for further information on what each parameter means.}
\item{The object optionally contains arrays of posterior distribution quantiles (\code{quantiles}) and means
(\code{means}), posterior probabilities (\code{post.prob}, see \code{report.devs} and \code{remove.trend}),
and maximum a posteriori parameter estimates (\code{MAPs}).}
}
All arrays' dimensions go in the order of iterations/diagnostics/quantiles, then parameters, then chains.
}
\description{
This function processes tree and trait data, runs a Stan-based Hamiltonian Monte Carlo (HMC) sampler to fit
these data to an evorates model, and returns the output of this sampler in a (relatively) user-friendly format.
}
\details{
PARAMETER DEFINITIONS HERE
}
\examples{


}
\concept{evorates fitting functions}
