% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUT_OPS_main.R
\name{\%chains\%}
\alias{\%chains\%}
\title{Extract posterior distribution samples from a fitted evorates model}
\usage{
fit \%chains\% select
}
\arguments{
\item{fit}{An object of class "\code{evorates_fit}" or "\code{param_block}"}

\item{select}{A list with two elements: 1) A character or numeric vector. If of class character, the given
text is matched to parameter
names using regular expressions. If of class numeric, the numbers are matched to indices of rate parameters
("\code{R_i}"), including the root ("\code{R_0}"). 2) A vector of integers specifying particular iterations
to extract from the chain. If unsupplied, all iterations will be extracted.}
}
\value{
A numeric vector, matrix, or 3D array of class "\code{param_block}". The dimensions will always go
in the order of iterations, then parameters, then chains, collapsing any dimensions of length 1 and storing
associated name information as attributes.
}
\description{
This operator extracts samples from the posterior distributions for particular parameters from the
output of \code{fit.evorates} or \code{output.evorates}
}
\details{

}
\examples{
#requires example fitted model object
#get rate posterior distributions
example.fit\%chains\%1:nrow(example.fit$call$tree$edge)
#',' behavior
example.fit\%chains\%'X2,X1_evocov'
#'dev' behavior
example.fit\%chains\%'.' #all parameters EXCEPT for rate deviation parameters
example.fit\%chains\%c('.','dev') #all parameters
#non-default iteration extraction
example.fit\%xchains\%list('R_0',c(23,45,52))


}
\concept{evorates_fit operators}
