% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_operators.R
\name{\%chains\%}
\alias{\%chains\%}
\title{Extract posterior distributions from a fitted Evolving Rates model}
\usage{
fit \%chains\% select
}
\arguments{
\item{fit}{An object of class "\code{evorates_fit}"}

\item{select}{A list with two elements: 1) A character or numeric vector. If of class character, the given
text is matched to parameter
names using regular expressions. If of class numeric, the numbers are matched to indices of rate parameters
("\code{R_i}"), including the root ("\code{R_0}"). Exhibits special behavior with regards to rate deviation
parameters ("\code{R_i_dev}") and commas ("\code{,}"); see details below. 2) A vector of integers specifying
particular iterations to extract from the chain. If unsupplied, all iterations will be extracted.}
}
\value{
A numeric vector, matrix, or 3D array. The dimensions will always go in the order of iterations,
then parameters, then chains, collapsing any dimensions of length 1 and storing associated name
information in attributes.
}
\description{
This is an operator for extracting sampled values for particular parameters from the output of \code{fit.evorates}
or \code{output.evorates}
}
\details{
These operators become really useful with some knowledge on how to use regular expressions.
Please consult Rstudio's
\href{https://rstudio.com/wp-content/uploads/2016/09/RegExCheatsheet.pdf}{Regular Expressions Cheat Sheet}
to learn more about regular expressions.

Since rate deviation parameters are oftentimes redundant with rate parameters, parameter names ending in
"\code{_dev}" are by default excluded unless text with "\code{dev}" is explicitly included at the end of a
text element the select argument.

(NOTE: THIS BEHAVIOR STILL EXISTS, BUT MULTIVARIATE EVORATES MODELS ARE DEPRECIATED FOR THE TIME BEING, SO NO
EVORATES MODELS ACTUALLY INCLUDE COVARIANCE MATRIX PARAMETERS...YET) For convenience, when faced with text
including a single comma ("\code{,}"), this operator will attempt to
switch the text preceding the comma with any text coming after the comma and before an underscore 
("\code{_}"). In practice, this helps extract parameters related to covariance matrices, as
"\code{trait1,trait2_evocov}" is automatically converted to "\code{trait2,trait1_evocov}". This is helpful
because corateBM model outputs only store values for the lower triangle of the covariance matrix, since
they are symmetric. Can return unpredictable results in cases where "\code{,}" or "\code{_}" are part of
trait names or tip labels. Use double-backslashes ("\code{\\\\,}" or "\code{\\\\_}") to indicate "escaped"
commas and underscores that the function should ignore during this swapping behavior. Note that you can
also use "\code{\\\\,}" to do encode regular expression searches of the form "\code{{m,n}}".
}
\examples{
#requires example fitted model object
#get rate posterior distributions
example.fit\%chains\%1:nrow(example.fit$call$tree$edge)
#',' behavior
example.fit\%chains\%'X2,X1_evocov'
#'dev' behavior
example.fit\%chains\%'.' #all parameters EXCEPT for rate deviation parameters
example.fit\%chains\%c('.','dev') #all parameters
#non-default iteration extraction
example.fit\%xchains\%list('R_0',c(23,45,52))


}
\seealso{
Other corateBM_fit operators: 
\code{\link{\%MAPs\%}()},
\code{\link{\%diagnostics\%}()},
\code{\link{\%means\%}()},
\code{\link{\%quantiles\%}()},
\code{\link{\%sampler\%}()}
}
\concept{corateBM_fit operators}
