% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OUT_OPS_main.R
\name{\%diagnostics\%}
\alias{\%diagnostics\%}
\title{Extract posterior distribution diagnostics from a fitted evorates model}
\usage{
fit \%diagnostics\% select
}
\description{
Extract posterior distribution diagnostics from a fitted evorates model
}
