% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corateBM_operator_utils.R
\name{\%diagnostics\%}
\alias{\%diagnostics\%}
\title{Extract posterior distribution diagnostics from a fitted correlated rates Brownian Motion model}
\usage{
fit \%diagnostics\% select
}
\arguments{
\item{fit}{An object of class "\code{corateBM_fit}", the output of a call to \code{fit.corateBM}.}

\item{select}{A list with two elements: 1) A character or numeric vector. If of class character, the given
text is matched to parameter names using regular expressions. If of class numeric, the numbers are matched
to indices of rate parameters ("\code{R_i}"), including the root ("\code{R_0}"). Exhibits special behavior
with regards to rate deviation parameters ("\code{R_i_dev}") and commas ("\code{,}"); see details below.
2) A character or numeric vector. If of class character, the given text is matched to diagnostic names
using regular expressions (with no special behavior). If of class numeric, the numbers are taken simply
taken as indices. The names of the available diagnostics are "\code{inits}", "\code{bulk_ess}", 
"\code{tail_ess}", and "\code{Rhat}", in that order. See details for definitions of what these diagnostics
mean. If only a vector (or list of length 1) is supplied, the operator assumes that vector corresponds to
the first element of the list and defaults to extracting all available diagnostics.}
}
\value{
A numeric vector, matrix, or 3D array. The dimensions will always go in the order of diagnostics,
then parameters, then chains, collapsing any dimensions of length 1 and storing associated name
information in attributes. When the output is of total length 1, defaults to being named according to
parameters and stores diagnostics and chains information as attributes.
}
\description{
This is an operator for efficiently extracting diagnostic summary statistics of posterior distributions for
particular parameters from output of the function \code{fit.corateBM}.
}
\details{
These operators become really useful with some knowledge on how to use regular expressions.
Please consult Rstudio's
\href{https://rstudio.com/wp-content/uploads/2016/09/RegExCheatsheet.pdf}{Regular Expressions Cheat Sheet}
to learn more about regular expressions.

For convenience, when faced with text including a single comma ("\code{,}"), this operator will attempt to
switch the text preceding the comma with any text comeing after the comma and before an underscore 
("\code{_}"). In practice, this helps extract parameters related to covariance matrices, as
"\code{trait1,trait2_evocov}" is automatically converted to "\code{trait2,trait1_evocov}". This is helpful
because corateBM model outputs only store values for the lower triangle of the covariance matrix, since
they are symmetric. Can return unpredictable results in cases where "\code{,}" or "\code{_}" are part of
trait names or tip labels. Use double-backslashes ("\code{\\\\,}" or "\code{\\\\_}") to indicate "escaped"
commas and underscores that the function should ignore during this swapping behavior. Note that you can
also use "\code{\\\\,}" to do encode regular expression searches of the form "\code{{m,n}}".

Since rate deviation parameters are oftentimes redundant with rate parameters, parameter names ending in
"\code{_dev}" are by default excluded unless text with "\code{dev}" is explicitly included at the end of
a text element the select argument.

Insert explanation of diagnostics here
}
\examples{
#requires example fitted model object and tree
#get all rate posterior distribution diagnostics
example.fit\%diagnostics\%1:nrow(example.tree$edge)
#',' behavior
example.fit\%diagnostics\%'X2,X1_evocov'
#'dev' behavior
example.fit\%diagnostics\%'.' #all parameters EXCEPT for rate deviation parameters
example.fit\%diagnostics\%c('.','dev') #all parameters
#specific diagnostics extraction
example.fit\%diagnostics\%list('R_0','0.432,'inits') #initial values
example.fit\%diagnostics\%list('R_0','0.432,'ess') #bulk and tail effective sample sizes


}
\seealso{
Other corateBM_fit operators: 
\code{\link{\%MAPs\%}()},
\code{\link{\%chains\%}()},
\code{\link{\%means\%}()},
\code{\link{\%quantiles\%}()},
\code{\link{\%sampler\%}()}
}
\concept{corateBM_fit operators}
