% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corateBM_operator_utils.R
\name{\%means\%}
\alias{\%means\%}
\title{Extract posterior distribution means from a fitted correlated rates Brownian Motion model}
\usage{
fit \%means\% select
}
\arguments{
\item{fit}{An object of class 'corateBM_fit', which is the typical output of a call to
\code{fit.corateBM}.}

\item{select}{A character or numeric vector. If of class character, the given text is matched to
parameter names using regular expressions. If of class numeric, the numbers are matched to indices
of rate parameters (\code{R_i}), including root (\code{R_0}). Exhibits special behavior with regards to
rate deviation parameters (\code{R_i_dev}) and commas (','); see details below.}
}
\value{
A numeric vector or matrix. The dimensions will always go in the order of parameters then chains,
collapsing any dimensions of length 1. When \code{select} matches to a single parameter, the parameter
name is stored as an attribute labelled "parameters".
}
\description{
This is an operator for efficiently extracting means of sampled parameter values for particular
parameters from output of the function \code{fit.corateBM}. Doesn't require a pre-stored means element.
}
\details{
These operators become really useful with some knowledge on how to use regular expressions.
Please consult Rstudio's
\href{https://rstudio.com/wp-content/uploads/2016/09/RegExCheatsheet.pdf}{Regular Expressions Cheat Sheet}
to learn more about regular expressions.

For convenience, when faced with text including a single comma (','), this operator will attempt to
switch the text preceding the comma with any text comeing after the comma and before an underscore
('_'). In practice, this helps extract parameters related to covariance matrices, as
'trait1,trait2_evocov' is automatically converted to 'trait2,trait1_evocov'. This is helpful because
corateBM model outputs only store values for the lower triangle of the covariance matrix, since they are
symmetric. Can return unpredictable results in cases where ',' or '_' are part of trait names or
tip labels. Use double-backslashes ('\\,'/'\\_') to indicate 'escaped' commas and underscores that the
function should ignore during this swapping behavior. Note that you can also use '\\,' to do encode
expression searches of the form '{m,n}'.

Since rate deviation parameters are oftentimes redundant with rate parameters, parameter names ending in
'_dev' are by default excluded unless text with 'dev' is explicitly included at the end of a text
element the select argument.
}
\examples{
#requires example fitted model object and tree
#get rate posterior distributions
example.fit\%means\%1:nrow(example.tree$edge)
#',' behavior
example.fit\%means\%'X2,X1_evocov'
#'dev' behavior
example.fit\%means\%'.' #all parameters EXCEPT for rate deviation parameters
example.fit\%means\%c('.','dev') #all parameters


}
\seealso{
\code{\link{\%chains\%}}, \code{\link{\%quantiles\%}}, \code{\link{\%MAPs\%}}, \code{\link{\%sampler\%}}
}
