% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corateBM_output_utils.R
\name{\%sampler\%}
\alias{\%sampler\%}
\title{Extract MCMC sampler parameters from a fitted correlated rates Brownian Motion model}
\usage{
fit \%sampler\% select
}
\arguments{
\item{fit}{An object of class "\code{corateBM_fit}", the output of a call to \code{fit.corateBM}.}

\item{select}{A character or numeric vector. If of class character, the given text is matched to parameter
names using regular expressions. If of class numeric, the numbers are taken simply taken as indices. See
details for names/order of available parameters.}
}
\value{
A numeric vector, matrix, or 3D array. The dimensions will always go in the order of quantiles,
then parameters, then chains, collapsing any dimensions of length 1 and storing associated name
information in attributes. Note that sampler parameters include warmup iterations. If output from the
\code{\%chains\%} operator includes \code{n} iterations, these will correspond to the last \code{n}
iterations included in output from the \code{\%sampler\%} operator.
}
\description{
This is an operator for efficiently extracting parameters used to tune the behavior of the stan-based
Markov chain Monte Carlo (MCMC) sampler from output of the function \code{fit.corateBM}. Helpful for
assessing problematic MCMC behavior.
}
\details{
The names of the available parameters are "\code{accept_stat__}", "\code{stepsize__}",
"\code{treedepth__}", "\code{n_leapfrog__}", "\code{divergent__}", "\code{energy__}", and "\code{lp__}",
in that order. See Stan documentation for more information on what these parameters mean.
}
\examples{
#requires example fitted model object
#get log posterior probability (+ some arbitrary constant) of iterations in mcmc sampler
example.fit\%sampler\%'lp'
#same result using numeric select
example.fit\%sampler\%7
#get all sampler parameters
example.fit\%sampler\%1:7


}
\seealso{
\code{\link[rstan]{check_hmc_diagnostics}}

Other corateBM_fit operators: 
\code{\link{\%MAPs\%}()},
\code{\link{\%chains\%}()},
\code{\link{\%diagnostics\%}()},
\code{\link{\%means\%}()},
\code{\link{\%quantiles\%}()}
}
\concept{corateBM_fit operators}
