% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_operators.R
\name{\%sampler\%}
\alias{\%sampler\%}
\title{Extract MCMC sampler parameters from a fitted correlated rates Brownian Motion model}
\usage{
fit \%sampler\% select
}
\arguments{
\item{fit}{An object of class "\code{corateBM_fit}", the output of a call to \code{fit.corateBM}.}

\item{select}{A character or numeric vector. If of class character, the given text is matched to parameter
names using regular expressions. If of class numeric, the numbers are taken simply taken as indices. See
details for names/order of available parameters.}
}
\value{
A numeric vector, matrix, or 3D array. The dimensions will always go in the order of quantiles,
then parameters, then chains, collapsing any dimensions of length 1 and storing associated name
information in attributes. Note that sampler parameters include warmup iterations. If output from the
\code{\%chains\%} operator includes \code{n} iterations, these will correspond to the last \code{n}
iterations included in output from the \code{\%sampler\%} operator.
}
\description{
This is an operator for efficiently extracting parameters used to tune the behavior of the Stan-based
Markov chain Monte Carlo (MCMC) sampler from \code{corateBM_fit} object, as well as prior probabilities
and likelihoods associated with samples. Helpful for assessing MCMC behavior and potentially calculating
marginal likelihoods, among other things.
}
\details{
The names of the available parameters are "\code{accept_stat__}", "\code{stepsize__}",
"\code{treedepth__}", "\code{n_leapfrog__}", "\code{divergent__}", "\code{energy__}", "\code{prior}",
"\code{lik}", and "\code{ost}", in that order. See Stan documentation for more information on what the
first 6 parameters mean. \code{prior} is the (ln) prior probability of all parameter estimates, while
\code{lik} is the (ln) likelihood of the data given the parameter estimates. Both of these are on the
scale of the transformed data, and therefore differ from the "true" (ln) prior probability and
likelihood by a constant. \code{post} is simply calculated by summing \code{prior} and
\code{lik.power * lik}. \code{post} is distinct from Stan's "\code{lp__}", which is calculated with
respect to a reparameterized sampling scale used by Stan internally.
}
\examples{
#requires example fitted model object
#get log posterior probability (+ some arbitrary constant) of iterations in mcmc sampler
example.fit\%sampler\%'lp'
#same result using numeric select
example.fit\%sampler\%7
#get all sampler parameters
example.fit\%sampler\%1:7


}
\seealso{
\code{\link[rstan]{check_hmc_diagnostics}}

Other corateBM_fit operators: 
\code{\link{\%MAPs\%}()},
\code{\link{\%chains\%}()},
\code{\link{\%diagnostics\%}()},
\code{\link{\%means\%}()},
\code{\link{\%quantiles\%}()}
}
\concept{corateBM_fit operators}
