% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ops.R
\name{operators}
\alias{operators}
\alias{\%chains\%}
\alias{\%quantiles\%}
\alias{\%means\%}
\title{New operators}
\usage{
fit \%chains\% select

fit \%quantiles\% select

fit \%means\% select
}
\description{
New operators
}
