% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIZ_prof.plot.R
\name{prof.plot}
\alias{prof.plot}
\title{Plot profiles of posterior samples from fitted}
\usage{
prof.plot(
  x,
  smooth = FALSE,
  p = 0.05,
  lower.quant = NULL,
  upper.quant = NULL,
  lower.cut = NULL,
  upper.cut = NULL,
  add.lines = NULL,
  quant.lines = TRUE,
  add = FALSE,
  make.legend = TRUE,
  ...,
  param.args = c("col", "border", "lines.col", "alpha", "border.alpha", "lines.alpha"),
  chain.args = c("angle", "density"),
  inpar.args = c("lines.lty", "lines.lwd")
)
}
\arguments{
\item{tree}{An object of class "\code{phylo}"}
}
\value{

}
\description{
This function processes tree and trait data, runs a Stan-based Hamiltonian Monte Carlo (HMC) sampler to fit
these data to an evorates model, and returns the output of this sampler in a (relatively) user-friendly format.
}
\details{
PARAMETER DEFINITIONS HERE
}
\examples{


}
\concept{evorates plotting functions}
